Adds missing weight field to Flood actor JSON encoding
This field, while not present in the template (an oversight that is
not worth fixing at this point), was being read for the overall
character weight value. Since it was nonexistent, this caused the
sheet to display `NaN` for the character's overall weight. This
defaults the weight to 0 kg to avoid this.