Fixes ranged weapon faction parsing and ammo group ToJSON
The reason the weapon postables were breaking for ranged weapons
was twofold:

1. Special ammo group values in the `ToJSON` instance were being
set to the `WeaponGroup` value, so were not producing a valid value
for that field.

2. We were attempting to parse the faction from the raw name, so
it was always producing a `Nothing` and thus not filling the
faction training correctly.

This has been tested and confirmed that items from the ranged
weapon compendium are posting correctly even when their sheet has
never been opened (which populates the fields with the default
select options at render). We're also rendering properly if posted
directly from the sheet while still in the compendium.