Uses ranged weapon raw base name for name and nickname parsing
This switches to using the raw ranged weapon base name to parse
name and nickname values, rather than the shared common name. This
accounts for differences in name based on fire mode, so multiple
items with the same name don't get created (and therefore replaced
when inserted into the `CompendiumMap`.