Generates entry IDs with Show instance
This uses a `Show` instance to generate a string from the entire
entry item for use in generating IDs. This increases the
likelihood that each ID will be unique, at the expense of being
able to guarantee ID consistency.

In actual fact, we couldn't *really* guarantee ID consistency - we
could only guarantee it when the item name didn't change, which is
admittedly rarer than any other piece of data on the item changing.

In the future (re: 100DOS system), we can introduce an injection
command to build a user-defined Foundry link which will help to
maintain link integrity.