Generates entry IDs with hashed name and label
This returns to using only the item name and compendium label. We
can once again guarantee the resiliency of the entry ID (barring a
change to the compendium or item name) by hashing the text produced
by joining the label and item names. This guarantees a unique `Int`
outcome, which can be used as the seed for pure pseudo-random number
generation, thus guaranteeing the same ID for the same combination
of compendium and item name.

As a result, we no longer need `Show` instances for the overwhelming
majority of types, so these instances have been removed.

During testing, we discovered a bug: treating a line that begins
with a double-quotes character (\") results in the exclusion of the
Forerunner equipment compendium entirely. This is because _all_
Forerunner equipment items have a comma in their name, so the CSV
document wraps the value in double-quotes to read it as a single
value. We were excluding this character because of the Bestiary
sheet, where some hidden lines have multi-line values in their
cells that begin with this character. More strict exclusion logic
will be required when it comes time to build the Bestiary compendium.