Adds Bestiary compendium
This builds a compendium for all Bestiary enemies in the game. Not
all values are able to be parsed perfectly, so this represents the
best compromise for getting the most amount of information correct
with the minimal amount of GM oversight. Mostly, the GM need only
concern themselves with Skill tests (which can be done as they
access them) and trainings.