Splits factioned entries back into faction-based compendia
Compendium packs have returned to being split up by faction, with
only Abilities and The Flood not having a faction (though, the
latter acts as its own faction). This was a community decision,
and will be substituted after the move to v11 with single-subject
compendium packs filled with faction-specific folders.