Adds Vehicle compendia
This introduces a compendium of Vehicle Actors per faction, as the
community requested. This is the last compendium before there is
complete coverage (though some other ideas might get implemented in
the future).

This does not include the vehicle weapons as of yet - that will be
done in the next commit.