Sets up factionless compendium packs for v11 migration
This makes an initial pass at setting up compendium packs that are
not split up by faction, instead adding a `filed` function to the
`CompendiumEntry` class that returns an optional folder to declare
for the entry. These folders should (hopefully) help to sort the
entries into faction-specific folders where applicable sometime in
the future.