Generates LevelDB compendium pack
Rather than generating a .db file for NeDB, we're now generating a
LevelDB database for each compendium pack. We can generate the DBs
and transfer them to Foundry, but they are not yet able to be
opened or read by Foundry. This is likely because item keys and
keyspaces are not yet set up correctly. This will be done in future
commits.