Blocks Vehicle Actors from being added to the combat tracker
Vehicle Actors do not have initiative - they do not act on their
own, with the exception of automated turrets which have specific
rules around how and when they act, which is outside the normal
flow of combat.

This removes them from combats by checking all incoming attempts to
create a combatant, and rejects (with a warning) any Vehicle Actors.
This fixes a bug where combats would break when attempting to sort
combatants on their AGI characteristic, which Vehicles do not have
(at least not in the same way that other Actors do).