Exposes some of the system API for macros
This exposes the `Chat`, `Common`, and `Dice` modules to users in
`game.mythic` so they can use those functions in modules. Worth
nothing is that these functions weren't necessarily designed with
macro support in mind, so a lot of them are going to take HTML
elements as arguments. These can be spoofed in some cases but it's
not ideal. Having a high-quality system API - with documentation -
should be a high priority in the 100DOS system implementation.