Adds system API documentation
This adds basic documentation around the functions exported from
the system API in the previous commit. Most of the documentation
is leaning on the existing JSDocs documentation that was already
written in each module. Additional details have been added to
explain how to duplicate the values required in an HTML element
argument for functions that require it.