Changes EMP to a text-based special rule
This is to allow the EMP special rule to use dice values as well
as integers. This was done because the EMP special rule itself has
text indicating that its default is 1D5 half-actions - while no
instances of the EMP special rule for a weapon or effect that I
could find actually used dice, this makes it possible for future
or homebrew items.