Fixes walker melee attacks failing if successful
The root cause of this was trying to read training proficiencies
around unarmed attacks off of the vehicle instead of the owner.
Because we only reach this point if the attack was successful, the
failure would only occur on what would be a hit - hence why any
attacks with only failures would still post properly.