Updates compendium packs to latest formatting
This makes a number of formatting and parsing corrections for the
latest changes to the Foundry compendium pack generation using
Roll20 compendium sheets. As a consequence of this, we've also
fixed the error where thrown weapons were parsed as ranged weapons
by adding in new weapon types to the `weaponDetailsMap`.