Adds Macro type and compendium pack
This adds a new `Macro` type of `FoundryData` to build compendium
packs for Foundry macros. To build the macros, we load in a single
file containing the raw JavaScript for each macro from a dedicated
directory. So far, only one macro (provided by Warpugss and his
viewers) has been added.

Worth noting was the additional casing around the result of the
`weaponDetailsMap` lookup that was added to provide (arguably)
better defaults for Secondary Attack weapon types from the ranged
weapon sheet.