Adds drag data for removing items from character sheets
Items can now be dragged off of Actor sheets - they will be
duplicated in the place to which they are dragged, be that the
Items sidebar or another Actor sheet. This has been tested for
the inventory, abilities, educations, weapon summary, and shield
granting item summary sections on all four character sheet types.