Defaults Flood NPCs to trained in everything
Flood Bestiary entries are now equipped with all equipment and
faction trainings by default. This fixes the issue where they were
not trained in their own biological weapons, and brings the parser
into compliance with the change to Flood training rules.

Technically, Flood creatures are only trained with what their host
was trained in, but in the majority of cases this means anything
they wouldn't be trained with will also not appear in the game,
so their having the training is inconsequential. Once a Gravemind
comes into the picture, they have all trainings anyway, so the
point becomes moot.